refacto: Delete unused API endpoints
