+++ backend/app/api/platform/models/__init__.py
@@ -1,5 +1,4 @@
 from app.api.v2.models import (
-    Eval,
     Event,
     EventDefinition,
     Events,
@@ -16,8 +15,6 @@ from app.api.v2.models import (
     TaskFlagRequest,
     Tasks,
     TaskUpdateRequest,
-    Test,
-    Tests,
     UserMetadata,
     Users,
     TaskHumanEvalRequest,
